search setting complete

+++ src/main/java/com/websathi/connectmeapp/model/business/Service.java
@@ -12,5 +12,5 @@ import lombok.Setter;
 public class Service {
     public String _id;
     public String name;
-    public int charges;
+    public Double charges;
 }
